kaspa and function fix

+++ utils/utils.py
@@ -23,9 +23,8 @@ def get_current_balance(coin):
     }
 
     response = requests.post(api_url, headers=headers, data=payload)
-    current_balance = response.json()['result'][coin]
-
-    allocated_balance = get_allocated_balance(coin)
+    current_balance = (float)(response.json()['result'][coin])
+    allocated_balance = (float)(get_allocated_balance(coin))
     return current_balance + allocated_balance
 
 def get_allocated_balance(coin):
